chore: slideshow
