+++ cypress/e2e_cucumber/common/index.js
@@ -2,7 +2,6 @@ import './before_each_visit_root.js'
 import './add_dashboard_items.js'
 import './confirm_discard_changes.js'
 import './exit_without_saving.js'
-import './add_a_FILTERTYPE_filter.js'
 import './choose_to_edit_dashboard.js'
 import './click_on_the_FILTERTYPE_filter_badge.js'
 import './exit_print_layout.js'
