--- cypress/e2e_cucumber/slideshow/index.js
@@ -1,1 +1,0 @@
-'../common/index.js'
